backend/services/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
numpy==1.24.3
scikit-learn==1.3.0
openai==1.54.0
replicate==0.25.0
Pillow==10.2.0
google-cloud-aiplatform==1.38.0
google-auth==2.25.0
gradio-client==0.10.1
cloudinary==1.36.0

# Dependencias para Avatar 3D
requests==2.31.0
pygltflib==1.16.1
trimesh==4.0.5

# NUEVAS DEPENDENCIAS v4.0
redis==5.0.1
celery==5.3.4
pybreaker==1.0.1
tenacity==8.2.3
prometheus-client==0.19.0